+++ docs/examples/disciplines/dynamic/plot_evaluation_time.py
@@ -29,7 +29,6 @@ makes this possible.
 from __future__ import annotations
 
 from matplotlib import pyplot as plt
-from numpy import array
 
 from gemseo_fmu.disciplines.dynamic_fmu_discipline import DynamicFMUDiscipline
 from gemseo_fmu.problems.fmu_files import get_fmu_file_path
@@ -80,7 +79,7 @@ default_2 = (discipline.time, discipline.local_data["y"])
 # and execute the discipline setting a simulation time of 0.3 seconds
 # with custom input values:
 discipline.set_next_execution(simulation_time=0.3)
-discipline.execute({"mass.m": array([1.5]), "spring.c": array([1050.0])})
+discipline.execute({"mass.m": 1.5, "spring.c": 1050.0})
 
 # %%
 # and store the results:
@@ -91,7 +90,7 @@ custom_1 = (discipline.time, discipline.local_data["y"])
 # we repeat the experiment until the final time
 # with the same custom input values:
 discipline.set_next_execution(restart=False)
-discipline.execute({"mass.m": array([1.5]), "spring.c": array([1050.0])})
+discipline.execute({"mass.m": 1.5, "spring.c": 1050.0})
 
 # %%
 # and store the results: